Prevent negative rates being passed to Barlow-Beeston equation which then has no solution and results in an exception.
The Barlow-Beeston equation only has a solution for positive source 
rates, but it is possible for source rates to be negative at the point 
that adjust_expectations is called. By checking for negative rates at an 
earlier point when calculating the likelihood, which always results in 
either returning  -inf or throwing an exception, this problem can be 
avoided.